factor out cssProject
