+++ package.json
@@ -52,9 +52,10 @@
     "ui/tournamentSchedule",
     "ui/tournamentCalendar",
     "ui/tree",
+    "ui/@build/cssProject",
+    "ui/@build/jsProject",
     "ui/@build/tsProject",
     "ui/@build/tsPlugins",
-    "ui/@build/jsProject",
     "ui/@types/lichess"
   ]
 }
