+++ ui/@build/cssProject/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@build/cssProject",
+  "version": "2.0.0",
+  "private": true,
+  "author": "Thibault Duplessis",
+  "license": "AGPL-3.0",
+  "main": "index.js",
+  "dependencies": {
+    "gulp": "^4",
+    "gulp-sass": "^4.0.2",
+    "gulp-sourcemaps": "^2.6.4",
+    "gulp-autoprefixer": "^6.0.0",
+    "gulp-rename": "^1.4.0"
+  }
+}
