+++ ui/challenge/gulpfile.js
@@ -1,2 +1,2 @@
 require('@build/tsProject')('LichessChallenge', 'lichess.challenge', __dirname);
-require('../gulp/cssProject.js')(__dirname);
+require('@build/cssProject')(__dirname);
