+++ ui/editor/gulpfile.js
@@ -1,3 +1,2 @@
 require('@build/jsProject')('LichessEditor', 'lichess.editor', __dirname);
-
-require('../gulp/cssProject.js')(__dirname);
+require('@build/cssProject')(__dirname);
