+++ ui/notify/gulpfile.js
@@ -1,3 +1,2 @@
 require('@build/tsProject')('LichessNotify', 'lichess.notify', __dirname);
-
-require('../gulp/cssProject.js')(__dirname);
+require('@build/cssProject')(__dirname);
