+++ ui/tournament/gulpfile.js
@@ -1,3 +1,2 @@
 require('@build/tsProject')('LichessTournament', 'lichess.tournament', __dirname);
-
-require('../gulp/cssProject.js')(__dirname);
+require('@build/cssProject')(__dirname);
